WIP: Better raw and viz tests
