+++ hcp/tests/test_viz.py
@@ -3,7 +3,7 @@ from numpy.testing import assert_equal
 import mne
 import hcp
 from hcp.tests import config as tconf
-from hcp.viz import make_hcp_bti_layout  # , plot_coregistration
+from hcp.viz import make_hcp_bti_layout, plot_coregistration
 
 hcp_params = dict(hcp_path=tconf.hcp_path,
                   subject=tconf.test_subject)
@@ -19,7 +19,6 @@ def test_make_layout():
 
 def test_plot_coregistration():
     """Test plotting of coregistration."""
-    # plot_coregistration(subjects_dir=tconf.subjects_dir)
-    pass
+    plot_coregistration(subjects_dir=tconf.subjects_dir, **hcp_params)
 
 mne.utils.run_tests_if_main()
